Refactoring orgy part two: * extracted read/reporting actions to their own modules: csv, log, report * fixed some edit bugs. * refactored some formatting * removed unnecessary interrupt functionality
